Move luigi.hadoop_jar to contrib
